Add Background class add to modules and link up
